+++ train.py
@@ -79,11 +79,11 @@ def train_dnn(df, i, epoch, batch=0):
     if arguments.lstm:
 
         print("[INFO] using LSTM layers")
-        x_train = x_train.reshape((arguments.batchSize / arguments.dnnBatchSize) * (1.0 - arguments.testSize), arguments.dnnBatchSize, x.shape[1])
-        y_train = y_train.reshape((arguments.batchSize / arguments.dnnBatchSize) * (1.0 - arguments.testSize), arguments.dnnBatchSize, y.shape[1])
+        x_train = x_train.reshape(int((arguments.batchSize / arguments.dnnBatchSize) * (1.0 - arguments.testSize)), arguments.dnnBatchSize, x.shape[1])
+        y_train = y_train.reshape(int((arguments.batchSize / arguments.dnnBatchSize) * (1.0 - arguments.testSize)), arguments.dnnBatchSize, y.shape[1])
 
-        x_test = x_test.reshape((arguments.batchSize / arguments.dnnBatchSize) * arguments.testSize, arguments.dnnBatchSize, x.shape[1])
-        y_test = y_test.reshape((arguments.batchSize / arguments.dnnBatchSize) * arguments.testSize, arguments.dnnBatchSize, y.shape[1])
+        x_test = x_test.reshape(int((arguments.batchSize / arguments.dnnBatchSize) * arguments.testSize), arguments.dnnBatchSize, x.shape[1])
+        y_test = y_test.reshape(int((arguments.batchSize / arguments.dnnBatchSize) * arguments.testSize), arguments.dnnBatchSize, y.shape[1])
         
         if arguments.debug:
             print("--------RESHAPED--------")
